Split walk-in team card models by concern
