+++ apps/web/src/features/walk-in/walk-in-team-card-onboarding-metrics.ts
@@ -4,24 +4,10 @@ import type {
 	DeveloperProject,
 	DeveloperSession,
 	DimensionAnalysisDataPoint,
-	WrappedSourceSplit,
 	WrappedV1,
 } from "@rudel/api-routes";
-import type { TeamPageMemberRow } from "@/features/team/use-team-page-data";
 import type { WalkInOnboardingMetrics } from "@/features/walk-in/walk-in-onboarding-types";
-import type { WalkInTeamMemberCardStatItem } from "@/features/walk-in/WalkInTeamMemberCard";
-import { MAX_ANALYTICS_DAYS } from "@/lib/analytics-date-range";
 import { formatCompactWholeNumber } from "@/lib/format";
-
-interface BuildResolvedTeamCardRowParams {
-	accountLabel: string;
-	debugProfileImageSrc: string;
-	developerDetails: DeveloperDetails | undefined;
-	sessionUserEmail: string | undefined;
-	sessionUserId: string | undefined;
-	sessionUserName: string | undefined;
-	teamMemberRows: readonly TeamPageMemberRow[];
-}
 
 interface BuildWalkInOnboardingMetricsParams {
 	commitBreakdown: readonly DimensionAnalysisDataPoint[] | undefined;
@@ -30,125 +16,6 @@ interface BuildWalkInOnboardingMetricsParams {
 	developerProjects: readonly DeveloperProject[] | undefined;
 	developerSessions: readonly DeveloperSession[] | undefined;
 	wrappedMetrics: WrappedV1["metrics"] | undefined;
-}
-
-export function buildResolvedTeamCardRow(
-	params: BuildResolvedTeamCardRowParams,
-): TeamPageMemberRow {
-	const {
-		accountLabel,
-		debugProfileImageSrc,
-		developerDetails,
-		sessionUserEmail,
-		sessionUserId,
-		sessionUserName,
-		teamMemberRows,
-	} = params;
-	const currentUserRow = findCurrentUserRow({
-		sessionUserEmail,
-		sessionUserId,
-		teamMemberRows,
-	});
-	const { displayName } = resolveTeamCardDisplayName({
-		accountLabel,
-		currentUserRow,
-		sessionUserEmail,
-		sessionUserName,
-	});
-	const email = currentUserRow?.email ?? sessionUserEmail ?? null;
-
-	if (developerDetails && sessionUserId) {
-		return {
-			activeDays: developerDetails.active_days,
-			cost: developerDetails.cost,
-			displayName,
-			email,
-			favoriteModel: developerDetails.favorite_model,
-			hasActivity:
-				developerDetails.total_sessions > 0 ||
-				developerDetails.active_days > 0 ||
-				developerDetails.total_tokens > 0,
-			imageUrl: debugProfileImageSrc,
-			inputTokens: developerDetails.input_tokens,
-			lastActiveDate: developerDetails.last_active_date,
-			outputTokens: developerDetails.output_tokens,
-			role: currentUserRow?.role ?? "Tracked collaborator",
-			totalSessions: developerDetails.total_sessions,
-			totalTokens: developerDetails.total_tokens,
-			userId: sessionUserId,
-		};
-	}
-
-	if (currentUserRow) {
-		return {
-			...currentUserRow,
-			imageUrl: debugProfileImageSrc,
-		};
-	}
-
-	return {
-		activeDays: 0,
-		cost: 0,
-		displayName,
-		email,
-		favoriteModel: null,
-		hasActivity: false,
-		imageUrl: debugProfileImageSrc,
-		inputTokens: 0,
-		lastActiveDate: null,
-		outputTokens: 0,
-		role: "Tracked collaborator",
-		totalSessions: 0,
-		totalTokens: 0,
-		userId: sessionUserId ?? "walk-in-preview",
-	};
-}
-
-export function buildWalkInStatItems(
-	row: TeamPageMemberRow,
-	distinctProjectCount: number,
-	sourceSplit: readonly WrappedSourceSplit[],
-): WalkInTeamMemberCardStatItem[] {
-	const normalizedSourceSplit = normalizeSourceSplit(sourceSplit);
-
-	return [
-		{
-			key: "codex-share",
-			title: `${normalizedSourceSplit.codexShare}% of wrapped sessions came from Codex`,
-			icon: "codex",
-			value: `${normalizedSourceSplit.codexShare}%`,
-		},
-		{
-			key: "claude-share",
-			title: `${normalizedSourceSplit.claudeShare}% of wrapped sessions came from Claude Code`,
-			icon: "claude",
-			value: `${normalizedSourceSplit.claudeShare}%`,
-		},
-		{
-			key: "sessions",
-			label: "SESS",
-			title: `${row.totalSessions.toLocaleString()} sessions`,
-			value: row.totalSessions.toLocaleString(),
-		},
-		{
-			key: "days",
-			label: "DAYS",
-			title: `${row.activeDays.toLocaleString()} active days across ${MAX_ANALYTICS_DAYS.toLocaleString()} tracked days`,
-			value: row.activeDays.toLocaleString(),
-		},
-		{
-			key: "tokens",
-			label: "TOK",
-			title: `${row.totalTokens.toLocaleString()} total tokens`,
-			value: formatCompactWholeNumber(row.totalTokens),
-		},
-		{
-			key: "repos",
-			label: "REPOS",
-			title: `${distinctProjectCount.toLocaleString()} distinct tracked projects`,
-			value: distinctProjectCount.toLocaleString(),
-		},
-	];
 }
 
 export function buildWalkInOnboardingMetrics(
@@ -243,136 +110,6 @@ export function buildWalkInOnboardingMetrics(
 		totalTokens:
 			wrappedMetrics?.total_tokens ?? developerDetails?.total_tokens ?? 0,
 	};
-}
-
-export function getWrappedArchetypeIndex(index: number, count: number) {
-	return ((index % count) + count) % count;
-}
-
-export function formatShareCardCreatedAt(date: Date) {
-	return new Intl.DateTimeFormat("en-US", {
-		month: "2-digit",
-		day: "2-digit",
-		year: "numeric",
-	}).format(date);
-}
-
-function findCurrentUserRow(input: {
-	sessionUserEmail: string | undefined;
-	sessionUserId: string | undefined;
-	teamMemberRows: readonly TeamPageMemberRow[];
-}) {
-	const { sessionUserEmail, sessionUserId, teamMemberRows } = input;
-	const normalizedSessionEmail = normalizeEmail(sessionUserEmail);
-
-	return teamMemberRows.find((row) => {
-		if (sessionUserId && row.userId === sessionUserId) {
-			return true;
-		}
-
-		return (
-			Boolean(normalizedSessionEmail) &&
-			normalizeEmail(row.email) === normalizedSessionEmail
-		);
-	});
-}
-
-function resolveTeamCardDisplayName(input: {
-	accountLabel: string;
-	currentUserRow: TeamPageMemberRow | undefined;
-	sessionUserEmail: string | undefined;
-	sessionUserName: string | undefined;
-}) {
-	const { accountLabel, currentUserRow, sessionUserEmail, sessionUserName } =
-		input;
-	const meaningfulSessionUserName = getMeaningfulDisplayName(sessionUserName);
-
-	if (meaningfulSessionUserName) {
-		return {
-			displayName: meaningfulSessionUserName,
-			source: "session.name",
-		} as const;
-	}
-
-	const emailHandle = getEmailHandle(sessionUserEmail);
-
-	if (emailHandle) {
-		return {
-			displayName: emailHandle,
-			source: "session.emailHandle",
-		} as const;
-	}
-
-	const meaningfulCurrentUserDisplayName = getMeaningfulDisplayName(
-		currentUserRow?.displayName,
-	);
-
-	if (meaningfulCurrentUserDisplayName) {
-		return {
-			displayName: meaningfulCurrentUserDisplayName,
-			source: "teamRow.displayName",
-		} as const;
-	}
-
-	return {
-		displayName: getFallbackTeamMemberDisplayName(accountLabel),
-		source: "accountLabelFallback",
-	} as const;
-}
-
-function getFallbackTeamMemberDisplayName(accountLabel: string): string {
-	if (accountLabel.includes("@")) {
-		return accountLabel.split("@")[0] || "User";
-	}
-
-	return accountLabel || "User";
-}
-
-function getEmailHandle(email: string | undefined) {
-	if (!email) {
-		return undefined;
-	}
-
-	const [emailHandle] = email.split("@");
-	return emailHandle?.trim() || undefined;
-}
-
-function normalizeEmail(email: string | null | undefined) {
-	return email?.trim().toLowerCase() || undefined;
-}
-
-function getMeaningfulDisplayName(value: string | undefined) {
-	const normalizedValue = value?.trim();
-
-	if (!normalizedValue) {
-		return undefined;
-	}
-
-	if (
-		normalizedValue.toLowerCase() === "operator" ||
-		normalizedValue.toLowerCase() === "unknown teammate"
-	) {
-		return undefined;
-	}
-
-	return normalizedValue;
-}
-
-function normalizeSourceSplit(sourceSplit: readonly WrappedSourceSplit[]) {
-	return {
-		claudeShare: Math.round(getSourceSharePercent(sourceSplit, "claude_code")),
-		codexShare: Math.round(getSourceSharePercent(sourceSplit, "codex")),
-	};
-}
-
-function getSourceSharePercent(
-	sourceSplit: readonly WrappedSourceSplit[],
-	source: WrappedSourceSplit["source"],
-) {
-	return (
-		sourceSplit.find((sourceEntry) => sourceEntry.source === source)
-			?.session_share_percent ?? 0
-	);
 }
 
 function formatWalkInLabel(value: string | undefined) {
